% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddt.R
\docType{data}
\name{ddt}
\alias{ddt}
\title{DDT Dataset}
\format{
A data frame with (number of rows) rows and (number of columns) variables:
\describe{
  \item{DDT}{Numeric variable indicating DDT levels.}
  \item{other_column}{Descriptions of other columns if present.}
}
}
\usage{
ddt
}
\description{
A dataset of DDT levels.
}
\keyword{datasets}
