% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{Bootstrap Confidence Interval}
\usage{
myboot2(iter = 10000, x, fun = mean, alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{Number of bootstrap iterations}

\item{x}{Input data vector}

\item{fun}{Function to apply (e.g., "mean", "median")}

\item{alpha}{Significance level for the confidence interval}

\item{cx}{Text size for annotations}

\item{...}{Additional arguments passed to the histogram function}
}
\value{
A list containing the confidence interval (`ci`), the function applied (`fun`), and the input data (`x`)
}
\description{
Bootstrap Confidence Interval
}
