% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mypvalue.R
\name{mypvalue}
\alias{mypvalue}
\title{Compute and Visualize P-value for a t-test}
\usage{
mypvalue(t0, xmax = 4, n = 20, alpha = 0.05)
}
\arguments{
\item{t0}{Numeric. The observed test statistic for which the p-value is calculated.}

\item{xmax}{Numeric. The maximum x-axis limit for plotting the t-distribution curve. Default is 4.}

\item{n}{Integer. The sample size, used to determine the degrees of freedom (df = n - 1). Default is 20.}

\item{alpha}{Numeric. The significance level for the hypothesis test. Default is 0.05.}
}
\value{
A list containing:
\item{q}{The critical t-value cutoff for the two-tailed test at the specified alpha level.}
\item{pvalue}{The two-tailed p-value for the observed test statistic `t0`.}
}
\description{
Calculates the p-value for a given test statistic in a t-distribution
with `n` degrees of freedom. Visualizes the t-distribution, highlighting
the critical regions and the observed test statistic.
}
\details{
importFrom("stats", "dt", "pt")
}
\examples{
# Example usage of mypvalue function
set.seed(55)
x1 <- rnorm(30, mean = 25, sd = 5)
tcalc <- (mean(x1) - 23) / (sd(x1) / sqrt(30))
tcalc
mypvalue(t0 = tcalc, n = 30)
}
