% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{Calculate Optimal Number of Tickets}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{Numeric. The number of seats on the flight.}

\item{gamma}{Numeric. The probability that the plane will be truly overbooked.}

\item{p}{Numeric. The probability of a "show" (the likelihood that a ticket holder will show up).}
}
\value{
A list containing:
\describe{
  \item{nd_discrete}{The optimal number of tickets for the discrete case.}
  \item{nc_normal}{The optimal number of tickets for the normal approximation.}
}
The function also produces two plots for visualizing the objective functions.
}
\description{
This function calculates the optimal number of tickets to sell for a flight based on
a binomial distribution and its normal approximation, considering the probability of
overbooking.
}
\examples{
ntickets(N = 200, gamma = 0.02, p = 0.95)

}
